% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay.R
\name{boxPointsGrid}
\alias{boxPointsGrid}
\title{Create a box point grid}
\usage{
boxPointsGrid(data, res, stepsize = NULL, min_pts = NULL)
}
\arguments{
\item{data}{a matrix or data frame containing a set of points}

\item{res}{the number of bins in which to divide the data on each axis.}

\item{stepsize}{the width of each bin. Alternative to `res` (`res = 1/stepsize`)}

\item{min_pts}{minimum number of points to be considered for drawing a box.}
}
\value{
A data frame containing the coordinates of boxes (x and y values) and
    whether or not they containt at least min_pts points (z value)
}
\description{
Creates a grid of square boxes around each point, flagging empty and non-empty boxes
}
\author{
Giuseppe D'Agostino
}
