% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_is_valid.R
\name{code_is_valid}
\alias{code_is_valid}
\title{is_code_valid}
\usage{
code_is_valid(code, appName)
}
\arguments{
\item{code}{Class code.}
}
\value{
Logical indicating whether the supplied code is valid.
}
\description{
Checks if class code is valid.
}
