% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay.R
\name{findHoles}
\alias{findHoles}
\title{Find holes}
\usage{
findHoles(ib_df)
}
\arguments{
\item{ib_df}{a data frame containing a series of isoband contours created by}
}
\value{
A data frame containing ordered coordinates for polygon vertices and
   three columns indicating whether they are a hole ("inner") or not ("outer"),
   to which cluster they belong, and a sub-clustering to allow \code{ggplot2}
   to draw them as holes.
}
\description{
Finds which polygons represent holes within a set of polygons
}
\author{
Giuseppe D'Agostino
}
