% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_is_valid.R
\name{get_access_token_temp}
\alias{get_access_token_temp}
\title{get_access_token_temp}
\usage{
get_access_token_temp()
}
\value{
The temporary access token.
}
\description{
Gets temporary firebase access token.
}
