% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay.R
\name{get_teacher_email}
\alias{get_teacher_email}
\title{get_teacher_email}
\usage{
get_teacher_email(session)
}
\arguments{
\item{session}{Shiny session.}
}
\value{
Teacher email
}
\description{
Gets teacher email
}
