% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay.R
\name{makeBoundingBox}
\alias{makeBoundingBox}
\title{Build a bounding box}
\usage{
makeBoundingBox(squares, stepsize = 1)
}
\arguments{
\item{squares}{output of \code{makeSquares}}

\item{stepsize}{numeric, the size of the step for the grid}
}
\value{
A data frame containing x and y coordinates for square vertices
   (with a value = 1) and the bounding box filled with value = 0. The box is
   larger than the set of vertices by \code{stepsize}.
}
\description{
Creates a bounding box around a subset of square vertices on a grid
}
