% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay.R
\name{makeContour}
\alias{makeContour}
\title{Make an isoband contour}
\usage{
makeContour(data, stepsize, minsize = 8, min_pts = 1)
}
\arguments{
\item{data}{a matrix or data frame containing a set of points}

\item{stepsize}{numeric, the spacing of the points}

\item{minsize}{numeric, the minimum size of the contour (in number of vertices)
to be retained. Default is 8 (2 squares)}

\item{min_pts}{numeric, the minimum number of points to be included in a square
for the contour to be drawn. Default is 1.}
}
\value{
A list of data frames containing the ordered set of coordinates to draw
    isoband polygons together with their clustering (disjointed polygon)
}
\description{
Creates an isoband around a cloud of points assigned to a square grid
}
\author{
Giuseppe D'Agostino
}
