% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay.R
\name{makeGrid}
\alias{makeGrid}
\title{Build a grid}
\usage{
makeGrid(data, stepsize)
}
\arguments{
\item{data}{A matrix or data.frame containing 2 columns corresponding to x
and y coordinates of points}

\item{stepsize}{numeric, the size of the step for the grid}
}
\value{
A data frame containing x and y coordinates for grid points
}
\description{
Creates a regularly spaced grid covering a set of points
}
\author{
Giuseppe D'Agostino
}
