% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay.R
\name{makeOverlay}
\alias{makeOverlay}
\title{Make overlay contour}
\usage{
makeOverlay(
  data,
  stepsize,
  minsize,
  min_pts = 2,
  offset_prop = 0.01,
  join_polys = TRUE,
  smooth = TRUE,
  smoothness = 3
)
}
\arguments{
\item{data}{a matrix ord data frame containing x and y coordinates of points}

\item{stepsize}{numeric, the relative step size used to build the square grid,
expressed as proportion of the maximum span of the range of \code{data}}

\item{minsize}{numeric, the minimum size of an overlay cluster}

\item{min_pts}{numeric, the minimum number of points to be enclosed}

\item{offset_prop}{numeric, the proportion by which polygons are offset
(inflated if outer, deflated if holes)}

\item{join_polys}{logical, should overlapping polygons be joined? Default is TRUE}

\item{smooth}{logical, perform kernel smoothing of polygons? Default is TRUE}

\item{smoothness}{numeric, smoothness parameter for kernel smoothing. Default is 3}
}
\value{
A data frame containing ordered coordinates for polygon vertices and
   three columns indicating whether they are a hole ("inner") or not ("outer"),
   to which cluster they belong, and a sub-clustering to allow \code{ggplot2}
   to draw them as holes.
}
\description{
Make an overlay contour by creating isobands of a square tessellation
}
