% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay.R
\name{makeSquares}
\alias{makeSquares}
\title{Build a grid of squares}
\usage{
makeSquares(grid, stepsize)
}
\arguments{
\item{grid}{output of \code{makeGrid}}

\item{stepsize}{numeric, the size of the step for the grid}
}
\value{
A data frame containing x and y coordinates for grid centers and for
    vertices of each square, named A, B, C, D and ordered clockwise from the
    top right one
}
\description{
Creates a grid of squares centered on an input grid
}
\author{
Giuseppe D'Agostino
}
