% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay-package.R
\docType{package}
\name{overlay-package}
\alias{overlay}
\alias{overlay-package}
\title{overlay: Render and Overlay ggplot2 and gt objects or Code Snippets with Perspective and Transparency}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Renders ggplot2 plots, gt tables, and carbon.js code images as magick objects, applies optional perspective warp distortions, and composites them over background images with transparency.
}
\author{
\strong{Maintainer}: Luis D. Verde Arregoitia \email{luis@liomys.mx} (\href{https://orcid.org/0000-0001-9520-6543}{ORCID}) [copyright holder]

}
\keyword{internal}
