% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay.R
\name{overlay}
\alias{overlay}
\title{overlay}
\usage{
overlay(input, output, session, duration = 90, forceClassCode = FALSE)
}
\arguments{
\item{input}{Shiny input.}

\item{output}{Shiny output.}

\item{session}{Shiny session.}

\item{duration}{Time in minutes until session expiry.}
}
\description{
Shrouds shiny app behind overlay.
}
