% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay.R
\name{poly_smooth}
\alias{poly_smooth}
\title{Smooth polygons}
\usage{
poly_smooth(poly, smoothness = 3, min_points = 8)
}
\arguments{
\item{poly}{a data frame containing ordered coordinates with polygon vertices}

\item{smoothness}{numeric, the extent of kernel smoothing. Higher means
rounder shapes. Default is 3.}

\item{min_points}{numeric, the minimum number of vertices to smooth.
Default is 8.}
}
\value{
A data frame containing ordered coordinates for polygon vertices and
   three columns indicating whether they are a hole ("inner") or not ("outer"),
   to which cluster they belong, and a sub-clustering to allow \code{ggplot2}
   to draw them as holes.
}
\description{
Uses kernel smoothing to smooth polygons
}
