% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_is_valid.R
\name{read_firebase}
\alias{read_firebase}
\title{read_firebase}
\usage{
read_firebase(path)
}
\arguments{
\item{path}{Which collection to read from, choose classCodes or launches}
}
\value{
Contents of firebase collection.
}
\description{
Reads from firebase.
}
