% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay.R
\name{teacher_code_is_valid}
\alias{teacher_code_is_valid}
\title{teacher_code_is_valid}
\usage{
teacher_code_is_valid()
}
\value{
Whether teacher code is valid.
}
\description{
Checks if teacher code is valid.
}
